Refactor CVSS 3 and below parsers to be more generic
