+++ src/Calculators/AbstractCvss3Calculator.php
@@ -4,15 +4,15 @@ declare(strict_types=1);
 
 namespace Rootshell\Cvss\Calculators;
 
-use Rootshell\Cvss\ValueObjects\CvssObject;
+use Rootshell\Cvss\ValueObjects\Cvss23Object;
 
 abstract class AbstractCvss3Calculator implements CvssCalculator
 {
-    abstract public function calculateModifiedImpactSubScore(CvssObject $cvssObject): float;
-    abstract public function calculateModifiedImpact(CvssObject $cvssObject): float;
+    abstract public function calculateModifiedImpactSubScore(Cvss23Object $cvssObject): float;
+    abstract public function calculateModifiedImpact(Cvss23Object $cvssObject): float;
     abstract public function roundUp(float $number): float;
 
-    public function calculateBaseScore(CvssObject $cvssObject): float
+    public function calculateBaseScore(Cvss23Object $cvssObject): float
     {
         $cvssObject->impactSubScore = $this->calculateImpactSubScore($cvssObject);
         $cvssObject->impact = $this->calculateImpact($cvssObject);
@@ -22,43 +22,43 @@ abstract class AbstractCvss3Calculator implements CvssCalculator
             return 0;
         }
 
-        if ($cvssObject->scope === CvssObject::SCOPE_UNCHANGED) {
+        if ($cvssObject->scope === Cvss23Object::SCOPE_UNCHANGED) {
             return $this->roundUp(min($cvssObject->impact + $cvssObject->exploitability, 10));
         }
 
         return $this->roundUp(min(1.08 * ($cvssObject->impact + $cvssObject->exploitability), 10));
     }
 
-    private function calculateImpactSubScore(CvssObject $cvssObject): float
+    private function calculateImpactSubScore(Cvss23Object $cvssObject): float
     {
         return 1 - ((1 - $cvssObject->confidentiality) * (1 - $cvssObject->integrity) * (1 - $cvssObject->availability));
     }
 
-    private function calculateImpact(CvssObject $cvssObject): float
+    private function calculateImpact(Cvss23Object $cvssObject): float
     {
-        if ($cvssObject->scope === CvssObject::SCOPE_UNCHANGED) {
+        if ($cvssObject->scope === Cvss23Object::SCOPE_UNCHANGED) {
             return 6.42 * $cvssObject->impactSubScore;
         }
 
         return 7.52 * ($cvssObject->impactSubScore - 0.029) - 3.25 * (($cvssObject->impactSubScore - 0.02) ** 15);
     }
 
-    private function calculateExploitability(CvssObject $cvssObject): float
+    private function calculateExploitability(Cvss23Object $cvssObject): float
     {
         return 8.22 * $cvssObject->attackVector * $cvssObject->attackComplexity * $cvssObject->privilegesRequired * $cvssObject->userInteraction;
     }
 
-    public function calculateTemporalScore(CvssObject $cvssObject): float
+    public function calculateTemporalScore(Cvss23Object $cvssObject): float
     {
         return $this->roundUp($cvssObject->baseScore * $cvssObject->exploitCodeMaturity * $cvssObject->remediationLevel * $cvssObject->reportConfidence);
     }
 
-    private function calculateModifiedExploitability(CvssObject $cvssObject): float
+    private function calculateModifiedExploitability(Cvss23Object $cvssObject): float
     {
         return 8.22 * $cvssObject->modifiedAttackVector * $cvssObject->modifiedAttackComplexity * $cvssObject->modifiedPrivilegesRequired * $cvssObject->modifiedUserInteraction;
     }
 
-    public function calculateEnvironmentalScore(CvssObject $cvssObject): float
+    public function calculateEnvironmentalScore(Cvss23Object $cvssObject): float
     {
         $cvssObject->modifiedImpactSubScore = $this->calculateModifiedImpactSubScore($cvssObject);
         $cvssObject->modifiedImpact = $this->calculateModifiedImpact($cvssObject);
@@ -68,7 +68,7 @@ abstract class AbstractCvss3Calculator implements CvssCalculator
             return 0;
         }
 
-        if ($cvssObject->modifiedScope === CvssObject::SCOPE_UNCHANGED) {
+        if ($cvssObject->modifiedScope === Cvss23Object::SCOPE_UNCHANGED) {
             return $this->roundUp(
                 $this->roundUp(
                     min($cvssObject->modifiedImpact + $cvssObject->modifiedExploitability, 10)
